delete post if user is authorized

+++ contollers/api/v1/posts_api.js
@@ -2,6 +2,7 @@ const Posts = require('../../../models/post');
 const Comments = require('../../../models/comments');
 module.exports.index = async (req, res) => {
     try {
+        console.log(req.user)
         var posts = await Posts.find({}).sort('-createdAt')
             .populate('user', '-password')
             .populate({
@@ -31,18 +32,24 @@ module.exports.index = async (req, res) => {
 module.exports.destroy = async (req, res) => {
     try {
         const post = await Posts.findById(req.params.id).exec();
-
-        await Comments.deleteMany({ post: req.params.id });
-        var deletePost = await Posts.findByIdAndDelete(post.id);
-        return res.json(200,{
-            message:"successfully deleted post and associated comments from db",
-            deletePost,
-        })
+        console.log(req.user.id,post);
+        if (post && post.user == req.user.id) {
+            await Comments.deleteMany({ post: req.params.id });
+            var deletePost = await Posts.findByIdAndDelete(post.id);
+            return res.json(200, {
+                message: "successfully deleted post and associated comments from db",
+                deletePost,
+            })
+        } else {
+            return res.status(401).json({
+                message: "you cant delete the post"
+            })
+        }
     } catch (error) {
 
         console.log('error while deleting post', error);
-        return res.json(500,{
-            message:"error in deleting",
+        return res.json(500, {
+            message: "error in deleting",
         })
     }
